2.17: changed corrosion debuff's info from 32-bit to 64-bit values

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Corrosion.java
@@ -89,12 +89,12 @@ public class Corrosion extends Buff implements Hero.Doom {
 
 	@Override
 	public String iconTextDisplay() {
-		return Integer.toString((int)damage);
+		return Long.toString((long) damage);
 	}
 
 	@Override
 	public String desc() {
-		return Messages.get(this, "desc", dispTurns(left), (int)damage);
+		return Messages.get(this, "desc", dispTurns(left), (long)damage);
 	}
 
 	@Override
